vsms-python-client-v1/requirements.txt
httplib2==0.10.3
urlextract
requests
cryptography
pyjwt